requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
typing-extensions>=4.0.0
keyring>=24.0.0

# Database
psycopg2-binary>=2.9.0

# LangChain & AI
langchain-openai>=0.1.0
langchain-experimental>=0.0.40
langchain_anthropic>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.40

# Vector Store & Embeddings
chromadb>=0.4.0
ollama>=0.1.0

# Financial Data
yfinance>=0.2.0
pandas>=2.0.0
stockstats>=0.6.0
eodhd>=1.0.0
finnhub-python>=2.4.0
akshare>=1.12.0
tushare>=1.2.0
backtrader>=1.9.0

# News & Social Media
praw>=7.7.0
feedparser>=6.0.0
parsel>=1.8.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0
pytz>=2023.3
typer>=0.9.0
rich>=13.0.0
questionary>=2.0.0

# Notifications
yagmail>=0.15.0
slack-sdk>=3.0.0
jinja2>=3.0.0

# Web & API
redis>=5.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
plotly>=5.18.0
chainlit>=1.0.0  # Chat UI for Eddie
websockets>=12.0  # WebSocket support for real-time audio
websocket-client>=1.6.0  # WebSocket client for testing

# Monitoring & Observability
langfuse>=2.0.0
prometheus-client>=0.20.0
prometheus-fastapi-instrumentator>=7.0.0
psutil>=5.9.0

# Voice (v2.0)
# TTS>=0.20.0  # Coqui TTS - Primary voice synthesis engine (requires Python <= 3.12)
# Note: TTS package doesn't support Python 3.13+. System will use pyttsx3 fallback.
# To install TTS, use Python 3.12 or earlier: python3.12 -m venv venv
pyttsx3>=2.90  # Fallback TTS - System TTS fallback
faster-whisper>=1.0.0  # Speech-to-Text (STT) - Low latency transcription
SpeechRecognition>=3.10.0  # Fallback STT engine

# Cognitive Architecture (v2.0)
networkx>=3.0  # Knowledge graph

# Web Crawling (v2.0)
crawl4ai>=0.3.0
duckduckgo-search>=4.0.0
beautifulsoup4>=4.12.0
playwright>=1.40.0  # For JavaScript rendering

# Build Tools
setuptools>=68.0.0
